Main Menu Update
*Created a separate class for the start game event, allowing the
instance to pass an argument to the gameplay class that contains an
array depicting which lessons were selected by the player, deck
generation will likely take place in the constrcutor of the gameplay
class.

*Added layers for sprites in Main such that the console is always
displayed on top of everything else in the game. This can be expanded in
the future if other areas of the game require such functionality.

*Added a placeholder background image to the main menu to make it look
pretty, shamelessly stolen from the internet, Nick can probably do the
art for our own image later.

*Added an FPS meter in the top right corner while the console is open.
I've also decided to switch to the dark theme for the minimal comps,
which has a nicer color scheme for the console window.

*Added tooltips to the Lesson Buttons on the main menu so that the user
knows which icons correspond to which lesson (if it wasn't already
obvious).

*Added a new MessageWindow class to relay information back to the user.