Groundwork for Deck Generation
*Basic random generation algorithm that inserts cards based on the
lessons chosen by the player, each card has equal weight in being chose,
those some cards are restricted from appearing more than once or twice.
Not sure if further bias against top-tier cards is necessary unless we
choose to implement difficulty levels that alter how the player's deck
is generated.

*Updated library.xml with the starting characters so that they may be
added to the deck during random generation. Also added additional tags
to each card for things like how many instances of that card are allowed
in each deck and which lesson is associated with each card in order for
random deck generation to get a list of valid cards that may be chosen
for the deck.

*Added additional functionality to the CardStack family of objects,
toString function prints each card only once and shows how many of that
card are in the stack (i.e. Charms x10), getNumCards returns how many
cards total are in the stack.

*Added additional functionality to the Deck class in order to prepare it
for random deck generation, the overwritten add function checks if the
card may be added to the deck via the maxAllowed property. setStarting()
will set the starting character to the given reference, currently no
checks are being made to verify that the argument is legitimate, but the
starting characters so far will be hard-coded anyway so it doesn't
matter.

*Made the console a little taller since I don't like scrolling very
much.