Small deck generation update
*Added different cases for 2 versus 3 lesson-type decks. All that's left
to do is incorporate adventure cards and extra characters (perhaps based
on secondary lesson types instead). After that, work will be have to be
diverted to adding cards to the sprite sheet and library file in order
to generate more unique and varied decks, as even though the deck
generation is currently random, each one is not very different since
there's isn't much choice in cards.

*After more cards are implemented, we can work on generating the deck
the AI will use against the player.