Big Deck Generation Update!
*Updated library.xml with all cards from the current sprite sheet
*All lessons types are now available for the player
*Implemented AI deck generation, the game now generates a random deck
for the AI opponent based on the lessons picked by the player, if the
player chooses to use creatures as a main lesson type, then so will the
AI, otherwise the AI will not pick creatures as their main lesson for
the sake of fairness.
*Fixed some parameter inconsistency with the addLesson function in
Gameplay.as
*Replaced harry_potter.utils.Console with fano.utils.Console for
consistency
*Uploaded the actual-size version of each card for easier description
look-up, still debating on whether to keep the simple tool tips or give
the player the option to view the full size version of the card during
gameplay.
*Actual gameplay is coming soon!