Tooltips + StatsPanel
--Main Changes--
Changed tooltips to use the proper card title in their descriptions, the
titles use a slightly larger font, which makes them stand out.
Began work on the stats panel for the player that shows several
information about the game. Current only number of lessons in play and
number of cards in the deck are tracked, when creatures are implemented
we'll keep track of the damage, and when actions are finally in we will
show the player how many moves he has left to do.

--Misc Changes --
Made the cards vector in CardStack protected again, added function
cardAt(index) which returns a card object at the given index.
TODO: Check bounds on the parameter.
Adjusted some positioning constants in the player class to make some
things look nicer.
Implemented fano.utils.DelayedFunctionCall for a nicer animation when
drawing the initial 7 card hand.
De-coupled the card titles from their descriptions in library.xml, this
was for the tooltip change that we made.