Beginning Creatures implementation
Laid out groundwork for creature cards, a check is made when a creature
card is played to see if the player has enough lessons to play it.

-Misc: Added a small glow effect when the player hovers over a clickable
card with their mouse, maybe do this with the deck as well.

TO DO:
*Implement discard pile for when certain creatures call for lessons to
be discarded in order to be played.

*Figure out positioning constants and offsets for creature cards and
discard pile